RR-2553 - Refactor DeletionReason for API-domain separation
Moves the `DeletionReason` enum from the `domain.entity` package to `resource.model` across Challenge, Condition, Strength, and Support Strategy controllers and services.

Introduces `DeletionReasonMapper` to explicitly convert between the `resource.model.DeletionReason` used by the API and the `domain.entity.DeletionReason` used internally by services. This improves architectural layering by preventing domain entities from being directly exposed in the API layer.
